backend_api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
numpy<1.24
pandas>=1.3.0,<2.0.0
prophet==1.1.5
google-cloud-tasks==2.16.1
google-cloud-storage==2.13.0
python-multipart==0.0.6
pytest==8.2.1
httpx==0.24.1
psycopg2-binary==2.9.9
